Do a first pass.
I totally revamped things. Now it will have pages of words, and I implemented
a template-based system for the text. I have to be careful, however, that it
a) works with python 2, which I'll test soon, and b) doesn't get caught up with
very short lists, which I'll test eventually...